requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==24.2.0
backoff==2.2.1
black==24.8.0
certifi==2024.8.30
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.7
distlib==0.3.8
distro==1.9.0
fastapi==0.115.0
filelock==3.16.1
frozenlist==1.4.1
grpcio==1.66.1
grpcio-tools==1.66.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.2
hyperframe==6.0.1
identify==2.6.1
idna==3.10
iniconfig==2.0.0
jiter==0.5.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.16
langchain-community==0.2.17
langchain-core==0.2.41
langchain-text-splitters==0.2.4
langsmith==0.1.125
marshmallow==3.22.0
mem0ai==0.1.15
monotonic==1.6
multidict==6.1.0
mypy==1.11.2
mypy-extensions==1.0.0
neo4j==5.24.0
nodeenv==1.9.1
numpy==1.26.4
openai==1.47.0
orjson==3.10.7
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
portalocker==2.10.1
posthog==3.6.6
pre-commit==3.8.0
protobuf==5.28.2
PyAudio==0.2.14
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
qdrant-client==1.11.2
rank-bm25==0.2.2
requests==2.32.3
simpleaudio==1.0.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
tenacity==8.5.0
tqdm==4.66.5
types-requests==2.32.0.20240914
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.30.6
virtualenv==20.26.6
yarl==1.11.1
